Change router_mac for T2 profile in test_interface script - PR to fix issue #4983 (#4995)
test_interfaces is failing for multi-asic chassis-packet system due to wrong router_mac. The test is picking the mac address of eth0 instead of picking the mac-address of the respective asic (in chassis-packet each easic has different router-mac). This is a PR to fix #4983. Fix is to use router-mac from config facts.