Change to fix issue #4955 - fix for multi-asic chassis_packet (#4997)
Proc is missing a check to see if a system is a multi-asic system or not and hence hitting issue #4955. This fix introduces a check to see if the system is multi-asic system or not and modifies the commands accordingly. This issue causes the test crm/test_crm.py::test_crm_fdb_entry to error out.
Co-authored-by: Perumal Venkatesh <pevenkat@cisco.com>