[link_flap] Log neighbor's bgp info just for specified neighbor type
What is the motivation for this PR?
Skip if the neighbor type is 'SmartCable', 'Server', 'Asic' and the
peer-tor in dual-tor system. Just focus on the neighbor which will
announce the BGP routes.

How did you do it?
This PR is to log the neigbor's bgp info while error happened for debug
purpose.

How did you verify/test it?
Run the cont link flap test case.

Signed-off-by: Kevin(Shengkai) Wang <shengkaiwang@microsoft.com>