Added changes for test_crm_neighbor to pass on Cisco-8000 (#4979)
What is the motivation for this PR?
Neighbor IP address (2.2.2.2) being configured in the existing testcase does not have route reachability so it is put in a pending list and not programmed in the hardware. Hence, it does not consume hardware resources and CRM neighbor available counter is not decremented. Adding the subnet (2.2.2.0/8) on the crm interface resolves this issue by providing reachability to the neighbor.

In this testcase, 1 percent of the available neighbor entries are added, i.e., 1 percent of 6.22k ~= 6k. In Sonic orchagent(neighorch) as a part of each neighbor addition, it creates SAI notification for creating a neighbor and a nexthop. The resource limit for nexthop being 4k, addition of 6k neighbors causes an orchagent crash. Hence, a reasonable number below the limit i.e., 3k neighbors are added in case of Cisco-8000 devices.

verify_thresholds() checks for logs when the threshold of 1 percent is reached, but because we have lowered the neighbors to less than the threshold, there will be no warnings. Thus, this check needs to be skipped.

How did you verify/test it?
Verified on cisco-8000 platform

Any platform specific information?
cisco-8000