Improve the KQL for creating Kusto tables/functions
The KQL statements for creating Kusto tables/functions have line wrapping issue.
Especially the statements for defining ingestion json mappings. They can't simply
be copied to the Kusto Explorer and run.

This change fixed the KQL syntax issue.

Another change is that the TestCases table is extended with an extra "Summary"
column. We need to pickup this change in the statements for creating tables.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>