Skip setting speed to 10G for T2 topology in test_config_interface_speed
test_config_interface_speed gets the current, and then sets the spped to 10G
and validates that the speed is set to 10G, and then sets it back to the original
speed.

However, in a 'T2' topology we are typically dealing with 100G/400G that don't support
setting the speed to 10G. If we sent such command, we are seeing orchagent crash.
So, for a T2 topology we skip setting and validating speed of 10G.