[renumber_topo] Set ipv6 route max size for ptf
What is the motivation for this PR?
The default ipv6 route max size is 4096, the kernel is continuously complaining the route cache is full with kernel version 5.4.0. After that, the ipv6 function will not work at all in the ptf. So we increase the value to 16800 which is the value used by eos. Should be no hurt with the old kernel version.

How did you do it?
sysctl -w net.ipv6.route.max_size=168000

How did you verify/test it?
Should not see the 'Route cache full' error in the output of dmesg.

Signed-off-by: Kevin(Shengkai) Wang <shengkaiwang@microsoft.com>