Shutdown all eBGP neighbors when running SFP platform tests (#4950)
* Shutdown all eBGP neighbors when running SFP platform tests

The SFP tests sequentially do negative actions on all the SFPs on all the DUTs. The only check is that the
command/API used worked - response is OK. Most of these tests will cause a quick link bounce (a few seconds)

In a scaled topology (like a T2 VOQ chassis) with 8K-12K routes per eBGP session, such quick link bounce
of all the links within a span of a minute causes a lot of stress on the CPU of the DUTs. A top on one of
the DUTs during the a test run on a linecard in a VoQ chassis shows that processes like 'zebra', 'bgpd',
'redis-server' 'orchargent' etc. are pegging at 50%-160% because of all the routes that have to be
deleted and propogated to all the remote iBGP peers. This causes the VoQ related tables to be in missing
some info. Specifically, we see that some eBGP neighbors are missing in the 'ip neighbor table' as well
as the ASIC_DB on the remote asics.

To avoid this stress, we shutdown eBGP on all the neighbors on all the DUTs before running the SFP related tests.
With eBGP disabled and thus all routes gone from eBGP, top shows that the processes are running at 2%-6% range as expected.

To accomplish this added a module scope fixture 'shutdown_ebgp' in duthost_utils that would:
- In setup:
  - get the number of total eBGP routes - using 'show ip route summary'
  - shutdown all eBgp neighbors
  - verify that the eBGP routes are 0
- In cleanup:
  - startup all eBGP neighbors
  - verify that the eBGP routes are the same what they were before we shutdown all eBGP neighbors

We then call this fixure in the SFP tests as needed.

* As part of shutdown_ebgp, check the orchangent CPU utilization (#4950 - sanmalho-git:shut_bgp)

Added check to make sure that the cpu utilization of orchagent after 'shutdown' or
'startup' of all eBGP neighbors is less than 10% for 5 consecutive seconds

* review comments

* Use a jitter of 5 in checking number of ebgp routes

Co-authored-by: Manouchehr Taheri <manouchehr.taheri@nokia.com>