Use one of the supported speeds to change the speed of the interface
In test_config_interface_speed, get the supported speeds on an interface.
If no speeds are supported or only the native speed is supported, then don't try to change
the speed. Otherwise use the first supported speed (other than the native speed) as the speed
to change on the interface, instead of the hard-coded 10G