Add reboot-cause history at the end of every reboot test to help for any failure analysis (#4836)
Description of PR
To improve the post sanity check and to help analyze any failure which might be caused by any reboot.
Should be same reboot-cause history for any test cases without any reboot involved for the test case.

What is the motivation for this PR?
To improve the post sanity check and to help analyze any failure which might be caused by any reboot.
Should be same reboot-cause history for any test cases without any reboot involved for the test case.

How did you do it?
I add an autouse fixture to check the uptime before and after the test running. If dut reboots during the test, then getting the reboot-cause history on it. 

How did you verify/test it?
By running any of the test cases.

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>