Merged PR 370: SCALE_ACL_FILE variable was removed by last revert, add it back
Code of scale cacl in test_cacl_appliation.py was revert due to block the access to testbed.
Enhanced code was summitted in github repo sonic-mgmt, it was merged into sonic-mgmt-int, but SCALE_ACL_FILE variable was missed in sonic-mgmt-int, because SCALE_ACL_FILE variable was removed by last revert.
Add it back.

Signed-off-by: Zhaohui Sun <zhaohuisun@microsoft.com>