Merged PR 375: Ignore acms in system health test
test_service_checker failed in nightly test.
It caused by “acms is not running”, it’s expected in STARLAB. We ignore acms service in test case config file.

ACMS (feature) is Microsoft only and is not relevant in our public Github repo. Hence the tests are also not relevant to the public repo. So commit it in internal repo.
ACMS runs in production but will not run in our STARLAB because STARLAB does not have connectivity to dSMS. Secondly, ACMS binary running inside ACMS container which is the heart of the container is already tested and validated by the dSMS team which own the binary. We might be able to test the other scripts running inside the ACMS container (owned by SONiC team) which do not depend on connectivity to dSMS, but we do not have a plan for that currently.

Signed-off-by: Zhaohui Sun <zhaohuisun@microsoft.com>