Merged PR 383: Add retry for locking testbed in nightly test pipeline template
Nightly test on some testbeds may need more than 24 hours. This caused the nightly test on next day may fail to lock the testbed because the previous nightly test is still running.

This change added retry logic for locking testbed. The timeout value is set to 2 hours. With this change, a nightly test job will retry locking testbed. If it failed to lock a testbed within 2 hours, then fail the job eventually.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>