Script to prepare the saiserver service
Use a script to prepare the saiserver serivce base on the syncd service.
The saiserver likes a light weight syncd, we can use the existing syncd
service manipulate script to control the saiserver.
Add a scipt to prepare the saiserver environment for saiserver.
Change the script name in the sai_qualify python

Testing done:
Test the generated script in DUT s6100

Signed-off-by: richardyu-ms <richard.yu@microsoft.com>