Skip cleanup docker images in the server cleanup script
One of the test server cleanup script is to delete all docker images.
This is not really necessary and may bite us in some scenarios. If
docker images are deleted, deploying a topology need to download the
docker image again. If something went wrong with docker registry,
deploying topology may be stuck.

This change removed the step of deleting docker images in the
server cleanup script.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>