Collect DB dump when test failed
What is the motivation for this PR?
Collect DB dump when test failed

How did you do it?
Add a common fixture collect_db_dump set as autouse. When test failed,
all of the DBs will be dumped, collected and sent to the test server

How did you verify/test it?
Failed a test and check the collection

Any platform specific information?
Supported testbed topology if it's a new test case?