[advance-reboot] Provide warmboot-SAD cases as a list argument (#5112)
What is the motivation for this PR?
To provide ability to selectively run SAD cases to avoid a full 15HR run at once.

How did you do it?
Provide SAD case list as an comma-separated input to argument sad_case_list. The inputs are case-insensitive.
White spaces are allowed between two testcases. Delimiter is still comma.
If the input SAD case is not accepted, it will be ignored and logic will still proceed with only acceptable list of case.

How did you verify/test it?
Executed tests on physical testbed with different SAD cases as input.