Update test report uploading to directly ingest into TestCases table (#5169)
What is the motivation for this PR?
Currently the test report uploading script firstly ingest test case results
into the RawTestCases table. Then an update policy automatically expands the
results in RawTestCases table and ingests into the TestCases table.

The problem is that each time when the uploading script is executed, it firstly
ingests all test results of the current round into a single row of the
RawTestCases table. The RawTestCases table has single column. This means that
all the test results are ingested into a single cell.

When the test results exceed certain size, it will hit the Kusto limitation of
single cell size. Eventually, data ingestion will fail silently.

How did you do it?
This PR introduced a new JSON mapping for the TestCases table. The uploading
script was also updated to directly ingest data into the TestCases table.
Ingested json will be mapped to columns according to the new mapping. A new
row is created for each json.

The RawTestCases table is no longer needed after this change. A future PR will
be created to remove the definition of the RawTestCases table after transition
to the new method is done.

This change is compatible with the current database schema. Just need to
ensure that the new mapping TestCasesMappingV1 is created in your database.

How did you verify/test it?
Tested on a standalone cluster with same tables, mappings and update policies.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>