Merged PR 411: Update the pipelines testing 202012 image to use internal-202012 branch
Update the pipelines testing 202012 image to use internal-202012 branch

We have created 202012 branch for the public and internal repo. The corresponding
branch in internal repo is called "internal-202012".

The pipeline yaml files for testing 202012 image need to be updated to use the new
internal-202012 branch.

This change is only for the internal-202012 branch.

After this PR is merged, the corresponding pipelines need to be updated to use
the yaml file from internal-202012 branch as well.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>

Cherry picked from !408