[advanced-reboot] New small disk SKU and handle bgpd.log between upgrades (#5218)
Previous PR (#5126) added small disk SKU handling for log_analyzer
However errors are seen for one SKU which was not added earlier. Moreover, auto detection mechanism is needed for 202012 upgrades, so that we don't just rely on hardcoded list (still needed for upgrade path).
In this PR:
Add missed SKU to the list - Arista-7060CX-32S-D48C8" - this list is needed for upgrade path with 201811 base.
Add detection of logs in tmpfs and automatically handle it. This case will still be not valid for upgrade path with 201811 base. But at least, it will cover 202012 upgrades, and will not require the hardcoded list to be updated for that path.
bgpd.log handling for SKUs with logs still present after upgrade - move quagga logs to frr dir so that ansible extract_log works fine.
Tested on small and large disk devices, and upgrade path proceeds fine.