Enhancing test_po_update_io_no_loss work on a multi-asic DUT (#4989)
* Making test_po_update_io_no_loss work on a multi-asic DUT

The test is enumerated to enum_frontend_asic_index, so will be
repeated per asic.

However, the test is picking up two portchannels randomly from all the asics
(using duthost.get_extended_mingraph_facts). The out_pc (portchannel on which
we are going to add and delete members) could be on an asic other than enum_frontend_asic_index.
In this case, the test fails.

Fix for this is to randomly pick out_pc from port channels that are on the same asic as
enum_frontend_asic_index. If we don't find a port channel in enum_frontend_asic_index, then
we would skip the test.

* Removing unused variable

* Change to pick out_pc on the same asichost was skipping for single asic

For singleasic, the namespace of asichost is None, while the namespace
in minigraph for the port channel members is '' (empty string).
Thus, we were failing to pickup any portchannels as out_pcs for single asic DUT.

* Rebasing and incorporating review comments.

- Pick out_pc from asichost mg_facts
- Removed restriction of having 2 ports in the all pcs to pick the in pc.

Co-authored-by: Manouchehr Taheri <manouchehr.taheri@nokia.com>