[unknown_mac] Add pkt string to the dict instead of the actual packet (#5272)
The following errors are seen in nightly run for this testcase. Though there isn't a repro available, this could be due to scapy version difference.
>           self.pkt_map[pkt] = (intf, get_intf_by_sub_intf(intf, vlan_id))
E           TypeError: unhashable type: 'Ether'

Hence instead of adding the pkt as a key, converting it into a string and using that as the dict key. This change might fix the issue

Fixes #5211
Signed-off-by: Neetha John <nejo@microsoft.com>

How did you verify/test it?
Ran the test with the changes and it passed