Fixing ptf port index in mg_facts for asics other than asic0 in a multi-asic linecard
In the minigraph_facts, minigraph_ptf_indices are derived by mapping this port_index of the device to device ports defined in the topology file that start with 0.
- PR #5011 introduced getting port index for ports on the device dynamically from the port_index defined in the port_config.ini file.
- PR #5223 fixed this to only doing so for a defined list of HWSKU's.

The minigraph_ptf_indices mapping for an asic (asichost.get_extended_minigraph_facts()) is wrong if we have a HWSKU like Nokia-IXR7250E-36x400G where
  - the port_index in the port_config.ini starts with 1
  - is a multi-asic linecard with Inband and Recycle Port - like Nokia-IXR7250E-36x400G.

- If we add this HWSKU to a the defined list of HWSKU to use port_config.ini, then the minigraph_ptf_indices mapping is off by 1 for asic0.
- If we don't add this HWSKU to a defined list of HWSKU's to use port_config_ini, then we revert to using port_utils.get_port_indices_for_asic method to get the device port index.
   - In this approach, we assume that the Inband and Recyle ports are frontend ports, and thus the starting port index for asic1 is off by 2.
     So, the minigraph_ptf_indices for asic1 is off by 2.

Because of the above, we will fail ipfwd/test_dip_sip.py, ipfwd/test_mtu.py and arp/test_arpall.py when run on any asic other than asic0 our such linecards

To fix this issue, we assume that the port index of the Inband and the Recyle ports are after all the front panel ports across the whole device.
Then we can treat the Inband and Recycle ports similar to backend ports

With the above fix, the minigraph_ptf_indices mapping is correct for all asics.