[multi-asic]: Fix port_alias to generate the right list of (#5261)
What is the motivation for this PR?
Provides fix for: #5141
This PR #5008 was added to ensure that the list of interfaces generated in port_alias matches with the list of interfaces generated in the show interfaces output. Show interfaces output currently displays both external and Inband/Recirc ports. So port_alias was modified to generate alias list for External/Inb/Rec ports if include_internal flag is set to false. This was causing an issue voq chassis multi-asic line card, when trying to access the dict with the list of alias to front panel asic interface names. Inb/Rec should not be present in front panel asic interfaces dict, so error was observed when trying to look for Inb/Rec port information in the front panel asic interfaces dict.

How did you do it?
Modified code to generate front panel asic interfaces list in sorted order or aliases.
Also, ensure that key(alias) is present before accessing front panel asic interfaces dict.

How did you verify/test it?
Able to generate minigraph for voq chassis multi-asic line card.
Change in multi-asic line card minigraph:
Ethernet-IB<> interfaces is getting added in the "EthernetInterfaces" section. Before this change Ethernet-IB<> interfaces was present only in SystemPort and VoqInbandInterfaces sections.