Fix DB dump with special characters in test names
In the case of a test failure, the redis DB is dumped to a file, tar'ed
up, and copied out. However, if there's characters in the test name (or
the file name) that have a special meaning in bash, then the command may
fail or behave unexpectedly. This is because the commands are currently
run in a shell.

Instead of doing this, have the commands be executed directly by
building an argv that can then be used to run the command. Because the
commands aren't going through a shell, there's no shell
expansion/parsing being done, and it's easier to specify arguments that
may have spaces or other special characters.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>