Merged PR 460: Schedule nightly test of n3163 testbeds to run earlier
Schedule nightly test of n3163 testbeds to run earlier

Nightly test of the n3164 testbeds are scheduled to 23:00 UTC time.
When the deep-clean test scheduled on Saturday 12:00 UTC is
triggered, the testing on n3164 testbeds usually are not completed.
This will block the deep-clean working on the server used by the
n3164 testbeds.

This change rescheduled the n3164 testbeds to run earlier.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>