[fdb] Enhance the stability of test_fdb
What is the motivation for this PR?
test_fdb is failed at multiple devices. It takes a long duration as it required sending lots of pkts between paired ports.
It is difficult to guarantee there is no congestion end-to-end between this period, especially at the server side as there are many workloads running on the server.

How did you do it?
To enhance the stability, if one pkt is not received at PTF, we can re-sent/verify this pkt for 3 times.

How did you verify/test it?
Run test_fdb.py

Signed-off-by: Kevin(Shengkai) Wang <shengkaiwang@microsoft.com>