[testcase]: MACsec LLDP test fixing for vSONiC (#5337)
Fix MACsec LLDP test cannot pass if the neighbor devices is vSONiC

What is the motivation for this PR?
The MAC address of SONiC host is locally administrated. So, LLDPd will use an arbitrary fixed value (00:60:08:69:97:ef) as the source MAC address of LLDP packet (https://lldpd.github.io/usage.html). But the MACsec driver in Linux used by SONiC VM has a bug that cannot handle the packet with different source MAC address to SCI if the send_sci = false.

How did you do it?
If send_sci = false and the neighbor device is SONiC VM, LLDPd need to use the real MAC address as the source MAC address.