Collect and archive PTF script logs and pcap files in nightly run
What is the motivation for this PR?
PTF script logs and pcap files were stored in PTF, which is not
convenient to get for analyse.

How did you do it?
Fetch PTF script logs and pcap files to tests/logs/.

How did you verify/test it?
Run a test.

Any platform specific information?
Supported testbed topology if it's a new test case?