Support both str and int type for slot_id received in pmon API
Based on PR https://github.com/Azure/SONiC/pull/945/, we should return
the sticker/label name on the chassis for the physical slot id
in the get_supervisor_slot PMON API and 'show chassis module status' command.

For Nokia linecards, the sticker label for supervisor is 'A'.
Thus we need to allow for string as possible return value as well - apart for int.