Merged PR 474: Add pipeline yaml for smoke test
Add pipeline yaml for smoke test

Currently we have PR test running on KVM testbed for new code
introduced to the sonic-mgmt repository. However, the coverage
of PR test is still not enough. Some changes may pass PR test
but fail the tests on physical testbed.

Purpose of this change is to introduce smoke test before merging
Github changes to our internal repository.

The smoke test pipeline has 2 stages. The first stage is to run
test on lab physical testbeds. The second stage is to merge the
Github changes to internal repository.

The test stage covers all the new changes to be merged and some
more tests that are not covered by PR testing. The code firstly
tries to find out all the test scripts updated by the incoming
changes. Run the touched scripts. Then it will run some more
tests that do not have the scripts all covered in PR testing and
some other scripts that are unstable or need too long to run.

After the testing is done, the results will be collected and
compared with average passing rate of the most recent runs on
same branch.

If the results of this run is not degraded over certain limit,
then we can consider the smoke test as passed. The analyzing
result is passed down to the next stage.

In the merge stage, if smoke test is passed or skipped, then
the new changes from Github can be merged.

The smoke test pipelines supports some run time parameters for
skipping the test stage or skipping uploading test results to
Kusto. Sometimes, we may can't wait for the tests to be
completed and just need to merge the changes from Github. Then
we can manually trigger the pipeline and tick the checkbox to
skip the test stage.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>