Support deployment of topology dualtor-mixed (#5405)
Approach
What is the motivation for this PR?
PR introduced the dualtor-mixed topology. This change is to support deployment of the new topology.

How did you do it?
Updated the playbook and ansible modules for deployment of the new dualtor-mixed topology.
Added vms-kvm-dual-mixed KVM testbed.
For the bridges created for active-active interfaces, the open flow rules are not added yet.
Need more PRs to address that after the details of the open-flow rules are sorted out.
How did you verify/test it?
Tested add-topo, remove-topo, restart-ptf for KVM dualtor-mixed testbed.
Tested add-topo, remove-topo, restart-ptf for physical dualtor-mixed testbed.