Add an option to do a safe config reload (#5374)
* Add an option to do a safe config reload

When doing a config reload, test cases cannot assume that the DUT is
available immediately after `config_reload` returns. Instead, they need
to wait on `duthost.critical_services_fully_started` to be true.

Therefore, in `config_reload` add an option `safe_reload` (defaults to
False to preserve current behavior) that will block on this so that the
function returns when the DUT is ready. In addition, have two test
cases (for now; other test cases will be checked later) use safe reload
so that the DUT is guaranteed to be ready when the function returns.

Otherwise, if the setup for the next test case is running while the
DUT is still bringing up the containers, then some check there could
fail.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>