Merged PR 484: Add pipeline for merging Github master to internal-stage branch
Add pipeline for merging Github master to internal-stage branch

The smoke test pipeline firtly try to merge the github change and run
tests with merged code. However, if there is merge conflicts, there is
no elegant way for the smoke test pipeline to resolve it.

This change added back the pipeline yaml for merging Github master
branch changes. The idea is to merge the change to a temporary branch
like `internal-stage`. Then the smoke test uses the internal-stage
branch for testing. After test passed, the smoke test then can push the
internal-stage branch to the internal branch.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>