Use dut_mg_facts instead of asic mg_facts to set in_ptf_index.
In a min T2 topology, we have one portchannel per asic.

However, test_po_update_io_no_loss requires 2 portchannels.

So, the test was enhanced to use the ingress portchannel from another asic.
But, the ptf index for this ingress portchannel was being derived from
the asic mg_facts of the egress portchannel on enum_frontend_asic_index which would not
have info about the ports in the other asic.

Fix was to use dut_mg_facts to get the ptf index for ingress portchannel on the other asic.