Fix loganalyzer no common config issue and re-balance PR testing (#5435)
What is the motivation for this PR?
Loganalyzer was broken in PR #3235. The issue is that common config was loaded
in subprocess for adding marks to syslog. After the subprocess exited, the
common config is lost.

PR #5193 tried to fix this issue. However, because of many new error logs sneaked
in when log analyzer was not working, PR testing started to fail by these error
logs after PR #5193 was merged.

PR #5391 and #5411 tried to work around the PR testing failure to unblock PR
testing. PR #5391 is to address the GCU related error logs and was merged.
PR #5411 tried to add other error logs to the common ignore list. But the effort
took too long because the ignore list seemed endless.

To unblock PR testing as soon as possible, the original fix #5193 was reverted
in #5433.

This PR tries to complete the work left over from #5411 and #5433.

How did you do it?
Changes:
* Fix the log analyzer common config not loaded issue.
* Temporarily add error logs to the common ignore list.
* Improve the logging of log analyzer and parallel_run
* PR testing t0_part2 takes much more time than t0_part1 after the GCU test
  scripts are added. This change re-balanced t0 part1&part2 testing by moving
  some of the tests from part2 to part1.
* Sorted the PR testing scripts in alphabetic order.

How did you verify/test it?
Tested run a few test scripts with log analyzer enabled on KVM testbed.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>