Merged PR 491: Update the smoke test pipeline to use stage branch
Originally, the smoke test pipeline firstly merge the change from Github and then run tests.
The problem is that if there are any conflicts, there is no way to pause the pipeline, manually resolve the conflicts and then resume tests. The pipeline will simply fail at the merging step.

The idea is that we still need a separate pipeline to do merge. Instead of directly push the merged content back to the same internal branch, we can push to a temporary branch, like internal-stage.

Then the smoke pipeline can start from the internal-stage branch. After test passed, push the internal-stage branch to the internal branch.

This change updated the code for merging and pushing.

Follow up PRs will be submitted to add separate pipelines for merging.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>