Merged PR 492: Refactor the merge pipeline to support pushing to stage branch
Refactor the merge pipeline to support pushing to stage branch

The existing code for merging change from github can only push to the
same base branch performed merge.

To support stage branch for somke test, we need to push the merged
changes to a temporary stage branch.

This change refactored the code for merge to support that.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>