Remove restriction on test_reboot.py
Missing hwsku.json might cause test_reboot.py to fail on some platforms
but not others. In previous commits, hwsku.json will be checked and if
missing, xfail the test. In this commit, the check is removed and some
machine where it is known will have the test skipped. In future commits,
this check will be implemented again but further limited to a few hwsku.

Signed-off-by: Xichen Lin <xichenlin@microsoft.com>