Fix pfc counters test for Onyx fanout
The test fails on testbeds with Onyx fanout switch, because the script pfc_gen.py is contained in a "storm" docker on Onyx fanout. The test was running pfc_gen.py directly from the Onyx switch, not from the docker.
Fix the commands run on Onyx fanouts to send PFC frames from the storm docker. Not changed for other fanout types.