[memory_checker] Checks Monit running status during teardown stage (#5575)
What is the motivation for this PR?
This PR aims to add a function which will continuously check whether Monit is running or not for 320 seconds after Monit was restarted at teardown stage. If Monit is not running, then this test case will be failed instead of failing the sanity check of next test case.

How did you do it?
A function is added to check the Monit running status at teardown stage.

How did you verify/test it?
I tested this change at lab device str-s6000-on-5.