Fix timeout logic for wr_arp ptf test (#5560)
Motivation for this PR:
In the HW SKUs with large number of ports, test_wr_arp is flaky. The present logic sends all the packets through all the ports, and then checks timeout. Inner-most loop can take ~20-25s in SKU with large number of ports.
This causes corner cases where the time has elapsed,but the logic still waits for 20s before checking timeout.

Change:
Move the timeout check from outer loop to inner loop. Additionally, fix the logs to indicate that the traffic thread has timed out. The present log say timedout on warmboot, which is not true.