Merged PR 541: Fix svc server name in inventory file
The SVC test server name in inventory file does not match the name in the connection graph file.

In inventory file, it was named as: `svcstr-serv-1`.
In connection graph file, it was named as `svcstr-server-1`.

Because of the mismatch, VLAN on root fanout switch was not properly configured during add-topo.
Consequently, all the traffic related testing failed for the testbeds using this server.

This fix updated the test server name in inventory file to match the actual hostname `svcstr-server-`.

Re-run add-topo, VLANs on root fanout can be properly configured for `svcstr-server-1`.