Fix the issue of finding test server IP address from inventory file (#5598)
What is the motivation for this PR?
Customized ansible module vmhost_server_info is for finding out IP
address of the test server based on server name.

This module is used by the playbook for configuring y_cable of dualtor topo.

This module assumes that all the test server hostname is defined as
vm_host_xx. This is not robust enough and may get wrong result under
some scenarios.

How did you do it?
This change improved robustness of this module. It firstly finds out all
the hosts belong to group specified by argument vmhost_server_name.
Then the first host which does not start with "VM" should be the test
server. This method is more robust.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>