[test gap] Add idle timeout test.  (#5599)
Description of PR
Linux-Unix shell variable TMOUT provides the “Auto Logout Functionality” for the login shell in case of no activity for a specified time. In this case, we will test if it can logout after the specified time in case of no activity.

What is the motivation for this PR?
Linux-Unix shell variable TMOUT provides the “Auto Logout Functionality” for the login shell in case of no activity for a specified time. In this case, we will test if it can logout after the specified time in case of no activity.

How did you do it?
The default TMOUT value of sonic is 900s, it's too long for us to test. So, firstly, modify the TMOUT value to 10s. Then, wait for more than 10s and observe whether the console session is logout.

How did you verify/test it?
Modify the TMOUT value to 10s and wait for more than 10s to observe whether the console session is logout.

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>