Comment out test cases which don't support KVM dualtor testbed。 (#5626)
Description of PR
In pr # 5388, there are lots of skipped testcases on kvm dualtor testbed, which case a long running time. These cases are under the folder tests/dualtor and tests/dualtor_io. Kvm testbed is the unsupported platform in these cases, so, comment out these cases to shorten the running time. When the kvm testbed is supported, we can delete the comments and running these cases.

What is the motivation for this PR?
In pr # 5388, there are lots of skipped testcases on kvm dualtor testbed, which case a long running time. These cases are under the folder tests/dualtor and tests/dualtor_io. Kvm testbed is the unsupported platform in these cases, so, comment out these cases to shorten the running time. When the kvm testbed is supported, we can delete the comments and running these cases.

How did you do it?
Comment out test cases which don't support the kvm dualtor testbed.

How did you verify/test it?
Running pipeline.

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>