Move out the skip check in `test_decap.py` to conditional mark. (#5641)
Description of PR
In previous version, the skip check was inside the script test_decap.py. We move it out to conditional mark. Also, we use the same 202012 sonic-mgmt branch for testing 201911 image, so 201911 should not be skipped in this case.

What is the motivation for this PR?
In previous version, the skip check was inside the script test_decap.py. We move it out to conditional mark. Also, we use the same 202012 sonic-mgmt branch for testing 201911 image, so 201911 should not be skipped in this case.

How did you do it?
Move out the skip check in test_decap.py to conditional mark and remove the skip oon MLNX for 201911 version.

How did you verify/test it?
Running on testbed which has 201911 version and some skipped platform.

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>