Merged PR 554: swap server 24 and 27
Server 24 keep crashing, suspect a driver issue. Swap server 24 with server 27 which is a new server.
When we say swap server, we want to exchange the hostname and IP info between these 2 servers, so we can keep minigraph and connection not changed but use a new server.
Several steps.
1. Exchange the IP.
In order to exchange the IP, we need to have 2 backup unused IPs when we do the swap, why it is 2, because currently on the server, there are 2 IP addresses for one port.
In this step, need to pay attention carefully, the IP access could lost due to un-expected mistake.
2. Exchange the hostname, follow below doc.
https://phoenixnap.com/kb/ubuntu-20-04-change-hostname#:~:text=1%20Use%20set-hostname%20to%20Change%20the%20Hostname%20Type,the%20result.%203%20Change%20the%20Pretty%20Hostname%20%28Optional%29
3. Exhange the pdu and root fanout connection info in below files as this PR does.
 ansible/files/sonic_str2_links.csv
 ansible/files/sonic_str2_pdu_links.csv
 ansible/files/spytest_connection_graph.xml