made TestbedProcessing to put IPv6 addresses for DUT and PTF to a lab file (#5504)
What is the motivation for this PR?
Enable tests which require DUT-PTF IPv6 connectivity, TACACS IPv6 on instance

There are TACACS IPv6 tests which communicate with TACACS server run in PTF via IPv6.
It is expected for addresses to be present in lab file to be eventually configured. Since lab file is generated by TestbedProcessing, we put addresses there and expect them to be generated into lab file but it does not happen. Fixed it

How did you do it?
Made TestbedProcessing to put IPv6 addresses for DUT and PTF to a lab file